requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.26.0
openai>=1.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
